Fix event listener

+++ command/src/app.ts
@@ -1,5 +1,4 @@
 import express from 'express';
-import { randomUUID } from 'crypto';
 import { OrderCreatedPublisher } from './events/order-created-publisher';
 import { OrderUpdatedPublisher } from './events/order-updated-publisher';
 import { OrderCancelledPublisher } from './events/order-cancelled-publisher';
@@ -13,34 +12,49 @@ app.post('/orders', async (req: express.Request, res: express.Response) => {
   const { name, productId, totalAmount } = req.body;
 
   const order = Order.build({
-    name, productId, totalAmount
-  })
+    name,
+    productId,
+    totalAmount,
+  });
   await order.save();
   console.log('COMMAND - ORDER CREATED:', {
     id: order.id,
     name: order.name,
     productId: order.productId,
-    totalAmount: totalAmount
+    totalAmount: totalAmount,
   });
 
   new OrderCreatedPublisher().publish({
     type: 'OrderCreated',
     data: {
-      id: randomUUID(),
-      name,
-      productId,
-      totalAmount,
+      id: order.id,
+      name: order.name,
+      productId: order.productId,
+      totalAmount: order.totalAmount,
     },
   });
 
   res
     .status(202)
-    .json({ message: 'Order created', data: { name, productId, totalAmount } });
+    .json({
+      message: 'Order created',
+      data: {
+        id: order.id,
+        name: order.name,
+        productId: order.productId,
+        totalAmount: order.totalAmount,
+      },
+    });
 });
 
 app.put('/orders/:id', async (req: express.Request, res: express.Response) => {
   const { id } = req.params;
   const { name, productId, totalAmount } = req.body;
+
+  if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+    res.status(400).json({ message: 'ID invalid' });
+    return;
+  }
 
   const order = await Order.findById(id);
   if (!order) {
@@ -49,15 +63,15 @@ app.put('/orders/:id', async (req: express.Request, res: express.Response) => {
   }
 
   order.set('name', name);
-  order.set('productId', productId)
-  order.set('totalAmount', totalAmount)
+  order.set('productId', productId);
+  order.set('totalAmount', totalAmount);
   await order.save();
 
   console.log('COMMAND - ORDER UPDATED:', {
     id: order.id,
     name: order.name,
     productId: order.productId,
-    totalAmount: totalAmount
+    totalAmount: totalAmount,
   });
 
   new OrderUpdatedPublisher().publish({
@@ -70,12 +84,15 @@ app.put('/orders/:id', async (req: express.Request, res: express.Response) => {
     },
   });
 
-  res
-    .status(202)
-    .json({
-      message: 'Order updated',
-      data: { id, name, productId, totalAmount },
-    });
+  res.status(202).json({
+    message: 'Order updated',
+    data: {
+      id: order.id,
+      name: order.name,
+      productId: order.productId,
+      totalAmount: order.totalAmount,
+    },
+  });
 });
 
 app.delete(
@@ -83,7 +100,12 @@ app.delete(
   async (req: express.Request, res: express.Response) => {
     const { id } = req.params;
 
-    await Order.findByIdAndDelete(id)
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      res.status(400).json({ message: 'ID invalid' });
+      return;
+    }
+
+    await Order.findByIdAndDelete(id);
 
     console.log('COMMAND - ORDER DELETED:', {
       id,
